WIP: Drop cleanup step in e2e tests
We have introduced a dependency on secrets that prevents the race
condition when kuttl deletes the namespace after running the test. We'll
thus leave ORC handle the proper ordering of resource deletion.

We however add a new step in the `create-minimal` test to ensure that
requests to delete the secret do not actually delete the secret thanks
to the finalizers.